IECoreGLPreview/LightVisualiser : adjustments after move
